# Remove source level query from javadoc hint.