plus small inner loop optimization in `TreeUtilities`.

The javadoc hint is typically in the top three list of slowest hints in files with many elements (fields, methods etc), even without the presence of javadoc sections.

Due to its high per-file invocation count, a source level or preferences query can end up being quite expensive. Luckily the source level query wasn't even used, Preferences will be addressed separately since this is riskier and influences all hints.

eliminates all purple sections:
![image](https://github.com/user-attachments/assets/0dcc99ab-85db-44e2-a0d8-445d3298f8e2)


this "after" measurement contains another optimization which is not included int his PR:

before:
```
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=9260.76ms, invocations=60007, cancelled=false}: error-in-javadoc
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=7213.45ms, invocations=10002, cancelled=false}: org.netbeans.modules.java.hints.bugs.NPECheck
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=4136.70ms, invocations=60007, cancelled=false}: org.netbeans.modules.java.hints.bugs.Unused
...
```


after:
```
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=6387.72ms, invocations=10002, cancelled=false}: org.netbeans.modules.java.hints.bugs.NPECheck
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=4796.10ms, invocations=60007, cancelled=false}: org.netbeans.modules.java.hints.bugs.Unused
FINE [org.netbeans.modules.java.hints.spiimpl.hints.HintsInvoker]: {time=1520.50ms, invocations=60007, cancelled=false}: error-in-javadoc
...
```
(note javadoc hint moved to place 3)

used https://github.com/mbien/nb-reprorepo/tree/master/performance/classes as test case